[REFACTOR] #33 - AsyncImage를 core.ui에서 core.designsystem으로 이동
