+++ feature/home/src/main/java/com/haman/feature/home/HomeScreen.kt
@@ -2,11 +2,9 @@ package com.haman.feature.home
 
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.haman.core.ui.item.AsyncImage
 
 @Composable
 fun HomeScreen(
     viewModel: HomeViewModel = hiltViewModel()
 ) {
-    AsyncImage(id = "0", load = viewModel::getImageByUrl)
 }
